+++ README.md
@@ -1,38 +1,64 @@
-## Learn More
-# +++ redux , manager page, nodejs, express
+# MUSIO
+#### https://creategreen.github.io/Musio/
+
+## Portfolio Site 
+제목: Musio 
+목적: 제가 진행했던 토이 프로젝트를 웹에서 보여주기 위해 개발하였습니다.
+기간: 2023/01 ~ 진행중 
+인원: 개인
+
+## Stack
+-언어
+    -JavaScript
+    -TypeScript
+-프레임워크
+    -React
+-툴킷/라이브러리
+    -GSAP
+    -THREE.js(R3F)
+
+## Preview with Code
 
 
 
-## Study webGL
-# https://blog.maximeheckel.com/posts/the-study-of-shaders-with-react-three-fiber/ 
-# https://codesandbox.io/s/03-adding-the-texture-refnp?from-embed=&file=/src/gl/glsl/vertex.glsl
-# https://tympanus.net/codrops/2020/03/17/create-a-wave-motion-effect-on-an-image-with-three-js/
-# https://docs.pmnd.rs/react-three-fiber/tutorials/loading-textures
-# https://velog.io/@kimbyungchan/react-threejs-shader-tutorial
-# https://heinleinsgame.tistory.com/8
 
-## ++study
-# svg animation
+## Trouble Shooting
+
+
+
+
+
+
+--------------
+#### References
+##### Study webGL
+https://blog.maximeheckel.com/posts/the-study-of-shaders-with-react-three-fiber/ 
+https://codesandbox.io/s/03-adding-the-texture-refnp?from-embed=&file=/src/gl/glsl/vertex.glsl
+https://tympanus.net/codrops/2020/03/17/create-a-wave-motion-effect-on-an-image-with-three-js/
+https://docs.pmnd.rs/react-three-fiber/tutorials/loading-textures
+https://velog.io/@kimbyungchan/react-threejs-shader-tutorial
+https://heinleinsgame.tistory.com/8
+
+##### svg animation
 https://ordinary-code.tistory.com/25 
 https://blog.avada.io/css/text-animations
 
-# hometext ex 
+##### hometext ex 
 https://codepen.io/tutsplus/embed/vdexgK?height=316&theme-id=0&default-tab=result 
 
-
-## R3F doc
+##### R3F doc
 https://docs.pmnd.rs/react-three-fiber/api/canvas
 https://docs.pmnd.rs/react-three-fiber/getting-started/examples
 https://docs.pmnd.rs/react-three-fiber/advanced/scaling-performance
 https://geekyants.com/blog/introduction-to-react-three-fiber/
 
-# three +ts shader bg
+##### three +ts shader bg
 https://dev.to/noprod/create-a-background-with-glsl-shader-threejs-typescript-i31
 
-# three+react road uniform texture
+##### three+react road uniform texture
 https://velog.io/@kimbyungchan/react-threejs-shader-tutorial
 
-# three.js background ex
+##### three.js background ex
 https://codepen.io/ykob/pen/BQmLeK
 https://www.awwwards.com/awwwards/collections/three-js/
 https://blog.maximeheckel.com/posts/vaporwave-3d-scene-with-threejs/
@@ -41,7 +67,7 @@ https://tympanus.net/codrops/2019/05/22/creating-grid-to-fullscreen-animations-w
 https://varun.ca/modular-webgl/
 https://blog.maximeheckel.com/posts/the-study-of-shaders-with-react-three-fiber/
 
-# shader to background error solve
+##### shader to background error solve
 https://github.com/pmndrs/react-three-fiber/issues/79
 https://meetup.nhncloud.com/posts/297
 https://tsh.io/blog/react-three-fiber/
