+++ public/index.html
@@ -15,7 +15,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/index.tsx"></script>
+   
     
   </body>
 </html>
