+++ src/components/Hometext.css
@@ -67,7 +67,7 @@
     
     padding:0;
     margin-left:auto;
-    margin-right:15px;
+    margin-right:50px;
     background: transparent;
     justify-content: center;
     align-items: center;
