+++ src/index.css
@@ -3,6 +3,6 @@ body{
     margin:0;
     overflow: hidden;
     min-height: 100vh;
-    
+    overflow-y: auto;
     
 }
