+++ src/index.tsx
@@ -8,11 +8,11 @@ const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
-  <React.Suspense fallback={null}>
-    
+  
+    <React.StrictMode >
       <App />
-    
-  </React.Suspense>
+      </React.StrictMode>
+  
 );
 
 // If you want to start measuring performance in your app, pass a function
