+++ src/routes/Project.js
@@ -130,7 +130,7 @@ const Project = () => {
       initial={{ opacity: 0 }}
       animate={{ opacity: 1 }}
       exit={{ opacity: 0 }}
-      transition={{ duration: 1 }}
+      transition={{ duration: 1.5 }}
     >
       <Canvas gl={{ antialias: false }} dpr={[1, 1.5]}>
         <ScrollControls
